vitaflow/competitor-analysis/myfitnesspal/requirements.txt
selenium>=4.0.0
Pillow>=9.0.0
requests>=2.28.0
